Refactor content script into multiple files
